refactor: dedupe single-command slash filter into common
The create-repo slash-command filter logic (input normalization plus
prefix matching) was copy-pasted between SlashCommandMenu and
useChatComposer. Extract filterSingleCommand and CREATE_REPO_COMMAND
into the shared common package and reuse from both call sites.
