+++ packages/common/src/index.ts
@@ -82,7 +82,9 @@ export {
   type SlashCommand,
   SLASH_COMMANDS,
   ABORT_COMMAND,
+  CREATE_REPO_COMMAND,
   filterSlashCommandsWithConflict,
+  filterSingleCommand,
 } from "./slash-commands"
 
 // Git operations
